The code for extraction of laboratory values v2
This code is used to extract lab values from CSV files that were exported from the hospital information system. The first version of the code was drafted in 2019 (LabExtract_V1_2019.py), and the second version of the cleaned code was drafted in 2021 (LabExtract_V2_2021.py). The third code, entitled "LabExtract_AllLabs_2022.py" is a code using the pivot function for extraction of all valid laboratory values and their interval from the admission (days).